src/backend/requirements.txt
astroid==3.3.11
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dill==0.4.1
astroid==3.3.11
blinker==1.9.0
click==8.1.8
dill==0.4.1
exceptiongroup==1.3.1
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.5
importlib_metadata==8.7.1
iniconfig==2.1.0
isort==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
iniconfig==2.3.0
isort==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
packaging==26.0
platformdirs==4.9.4
pluggy==1.6.0
Pygments==2.19.2
packaging==26.0
platformdirs==4.4.0
pluggy==1.6.0
Pygments==2.19.2
pylint==3.3.9
PyMySQL==1.1.2
PyMySQL==1.1.1
pytest==8.4.2
python-dotenv==1.2.2
SQLAlchemy==2.0.48
tomlkit==0.14.0
typing_extensions==4.15.0
Werkzeug==3.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
SQLAlchemy==2.0.48
tomli==2.4.0
tomlkit==0.14.0
typing_extensions==4.15.0
Werkzeug==3.1.6
zipp==3.23.0
